Fix statistics with spurious activations
It might happen (and happens!) that the kernel decides to block a task,
and then wakes up the task before returnin to user space. So, there is
no effect on the task, but a wake_up event is generated in ftrace (since
the "block" event is a context switch, it is not actually generated).
So, do not consider such spurious wake_up events when computing statistics.